devfs: Split devfs-random into two files using submodules
