+++ sys/fs/devfs/Makefile
@@ -1,2 +1,4 @@
 MODULE=devfs
+SRC = devfs.c auto_init_devfs.c
+SUBMODULES = 1
 include $(RIOTBASE)/Makefile.base
